C86 Separate out data; naming
