+++ src/constants.ts
@@ -6,5 +6,5 @@ export const WATERSHED_PMTILES_URL =
   'https://gpw-coastal-pollution-model-data-public-0001.s3.ap-southeast-2.amazonaws.com/app/watersheds/watersheds.pmtiles'
 export const GLOBAL_LULC_PMTILES_URL =
   'https://gpw-coastal-pollution-model-data-public-0001.s3.ap-southeast-2.amazonaws.com/outputs/land/regional/Central_Indo_Pacific_LULC_SDR.pmtiles'
-export const LULC_2000_TILES_URL =
+export const LULC_2000_URL =
   'https://gpw-coastal-pollution-model-data-public-0001.s3.ap-southeast-2.amazonaws.com/app/global_lulc/lulc_2000_visual.tif'
